\name{AffyGEx-package}
\alias{AffyGEx-package}
\alias{AffyPipelineHG2ST}
\docType{package}
\title{
\packageTitle{AffyGEx}
}
\description{
\packageDescription{AffyGEx}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AffyGEx}
\packageIndices{AffyGEx}
}
\author{
\packageAuthor{AffyGEx}

Maintainer: \packageMaintainer{AffyGEx}
}
