\name{gen_rma_FUN}
\alias{gen_nc}
\title{
Generating normalized and filtered rma object for Affymetrix gene and exon human arrays
}
\description{
This function takes as input objects generated by read.celfiles of the oligo package. The probeset information are added to feature data using getNetAffx, which include assigning single gene symbols per probe(set) and corresponding ensemble transcript IDs. Followed by a filtering step, of probes that have signals within specific standard devians from the average signal of negative control probes per chip. Only probes that are present within a predefined subset of chips will be retained. All control and unannotated probes are filtered and then probesets are reduced to gene set level retaining the median value of probesets per gene.
}
\usage{
gen_rma_FUN(raw_set, sd_val = 0, snames, ct_off, rds_path)
}
\arguments{
  \item{raw_set}{
 an object generated by read.celfiles
}
  \item{sd_val}{
a numeric vlaue of the fraction/number of standard deviations from the average signal of negative control probes, below which a probe is filtered out.
}
  \item{snames}{
sample names (must be identical to that of the raw_set object)
}
  \item{ct_off}{
a value between 0 and 1. This value allows retaining probes represented in 1-ct_off*100 samples
}
 \item{rds_path}{
a character string of the file name (which could include the path) to save the normalized data as an rds file
}
}
\details{
This function generates a normalized and filtered rma object
}
\author{
Ahmed Sadik
}

