\name{geo_df_fun}
\alias{geo_df_fun}
\title{
Parse a table of studies from a GEO database search of series.
}
\description{
This function extracts information from GEO series searches and returns them in a tabular form suitable for text mining.
}
\usage{
geo_df_fun(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
path to the geo search text file
}
}

\author{
Ahmed Sadik
}