\name{gn_sel_FUN}
\alias{gn_sel_FUN}
\title{
Gene selection for topGo analysis
}
\description{
This function allows filtering genes based on logFC before entering topGo
}
\usage{
gn_sel_FUN(score, thrsh_val)
}
\arguments{
  \item{score}{
logFC values
}
  \item{thrsh_val}{
filtering threshold
}
}

\author{
Ahmed Sadik
}