\name{TF_camera_gsa_FUN}
\alias{TF_camera}

\title{
Transcription factor based Gene Set Analysis
}
\description{
Perform Gene Set Analysis using camera of the limma package for signature lists derived from the transcription factor specific toptables.
}
\usage{
TF_camera_gsa_FUN(TF_list, top_T = NULL, rma_obj, pltfrm = "Affy", d_m, cntrst = NULL, cnt_mat = NULL, TF_motifs = FALSE, wts = NULL, res_path)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{TF_list}{
character vector of transcription factors (either combined or motifs)
}
  \item{top_T}{
topTable generates after applying hg38_annot_FUN
}
  \item{rma_obj}{
object generated by oligo::rma or an MAList
}
  \item{pltfrm}{
either affy or agilent
}
  \item{d_m}{
design matrix}
  \item{cntrst}{
the coeffecient to report for the camera output
}
  \item{cnt_mat}{
contrast matrix, not used}
  \item{TF_motifs}{
logical, defaults to FALSE if no transcription factor motifs are used
}
  \item{wts}{
a vector of weights applied when running camera, defaults to NULL
}
  \item{res_path}{
path to save GSA files
}
}

\author{
Ahmed Sadik
}
