\name{Vs_FUN}
\alias{Vs_FUN}
\title{Remove Vs from the file names of transcription factor target files
}
\description{
Remove Vs from the file names of transcription factor target files
}
\usage{
Vs_FUN(x)
}
\arguments{
  \item{x}{
character vector of file names of transcription factors in the GBMSygnal directory
}
}
\author{
Ahmed Sadik
}
